Web Data Access (MyRestController, MyWordService, WordRepository)
- WordRepository class now has a selectByLanguage method, allowing the
user to select words by a certain language. This method uses a specially
made query that is translated into the correct SQL for each database
vendor automatically. Should look into pagination in the future.

- WordService class is now an interface.

- MyWordService is an implementation of the WordService class.

- WordSource and Symbols no long contain a reference to a Word, the
references were causing infinite listings of a word which referred to a
source or symbol that referred to a word that... No issues have been
caused by this change so it should be permanent.

- MyRestController now allows access to getAllWords and
getWordsByLanguage (selectByLanguage) via get requests.